Move where we create the resolvers for our fields
