+++ strawberry/schema/schema_converter.py
@@ -1,4 +1,5 @@
 from enum import Enum
+from inspect import isasyncgen, iscoroutine
 from typing import Any, Callable, Dict, Type, cast
 
 from graphql import (
@@ -16,6 +17,7 @@ from graphql import (
     GraphQLNullableType,
     GraphQLObjectType,
     GraphQLOutputType,
+    GraphQLResolveInfo,
     GraphQLScalarType,
     GraphQLType,
     GraphQLUnionType,
@@ -27,6 +29,7 @@ from strawberry.directive import DirectiveDefinition
 from strawberry.enum import EnumDefinition, EnumValue
 from strawberry.field import StrawberryField
 from strawberry.scalars import is_scalar
+from strawberry.types.info import Info
 from strawberry.types.types import TypeDefinition, undefined
 from strawberry.union import StrawberryUnion
 
@@ -309,7 +312,55 @@ class GraphQLCoreConverter:
         return graphql_object_type
 
     def from_resolver(self, field: StrawberryField) -> Callable:
-        return field.get_wrapped_resolver()
+        def _check_permissions(source: Any, info: Info, kwargs: Dict[str, Any]):
+            """
+            Checks if the permission should be accepted and
+            raises an exception if not
+            """
+            for permission_class in field.permission_classes:
+                permission = permission_class()
+
+                if not permission.has_permission(source, info, **kwargs):
+                    message = getattr(permission, "message", None)
+                    raise PermissionError(message)
+
+        def _strawberry_info_from_graphql(info: GraphQLResolveInfo) -> Info:
+            return Info(
+                field_name=info.field_name,
+                field_nodes=info.field_nodes,
+                context=info.context,
+                root_value=info.root_value,
+                variable_values=info.variable_values,
+                return_type=field._get_return_type(),
+                operation=info.operation,
+                path=info.path,
+            )
+
+        def _resolver(_source: Any, info: GraphQLResolveInfo, **kwargs):
+            strawberry_info = _strawberry_info_from_graphql(info)
+            _check_permissions(_source, strawberry_info, kwargs)
+
+            result = field.get_result(_source, info=strawberry_info, kwargs=kwargs)
+
+            if isasyncgen(result):
+
+                async def yield_results(results):
+                    async for value in results:
+                        yield value
+
+                return yield_results(result)
+
+            if iscoroutine(result):  # pragma: no cover
+
+                async def await_result(result):
+                    return await result
+
+                return await_result(result)
+
+            return result
+
+        _resolver._is_default = not field.base_resolver  # type: ignore
+        return _resolver
 
     def from_scalar(self, scalar: Type) -> GraphQLScalarType:
         return get_scalar_type(scalar, self.type_map)
